requirements.txt
Django==4.2.16
django-appconf==1.1.0
django-compressor==4.5.1
django-multi-captcha-admin==2.0.0
django-ranged-response==0.2.0
django-recaptcha2==1.4.1
django-simple-captcha==0.6.2
django-summernote==0.8.20.0
django-robots==6.1
django-taggit==6.1.0
idna==3.10
mysqlclient==2.2.7
pillow==11.3.0
psycopg2-binary==2.9.10
python-decouple==3.8
rcssmin==1.1.2
requests==2.32.5
rjsmin==1.2.2
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
webencodings==0.5.1
